+++ src/lib/manager/index.ts
@@ -1,3 +1,2 @@
 export * from "./AutoHandler"
 export * from "./CommandHandler"
-