+++ src/preload/commands/restrict/index.ts
@@ -7,32 +7,7 @@ export default class Restrict extends Command {
         super(
             "restrict",
             "Restricts a command from being used in a channel. Using `{all}` as a command argument restrics all commands and using `#{all}` as a channel argument restricts the commands in all text channels.",
-            [
-                {
-                    name: "command",
-                    description: "A command to restrict. Can be \"{all}\" to restrict all commands.",
-                    required: true,
-                    type: "string"
-                },
-                {
-                    name: "command 2",
-                    description: "More commands to restrict.",
-                    required: false,
-                    type: "string"
-                },
-                {
-                    name: "channel mention",
-                    description: "The text channel to restrict commands in. Can be \"#{all}\" to restrict commands in all text channels.",
-                    required: false,
-                    type: "string"
-                },
-                {
-                    name: "channel mention 2",
-                    description: "More text channels to restrict commands in.",
-                    required: false,
-                    type: "string"
-                }
-            ],
+            undefined,
             undefined,
             ["SEND_MESSAGES", "MANAGE_CHANNELS"],
             ["text"]
@@ -40,7 +15,7 @@ export default class Restrict extends Command {
     }
 
     async run(client: Client, msg: Message): Promise<boolean> {
-        const args = await this.parseArgs(msg) as string[] | false;
+        const args: any = await this.getArgs(msg)
 
         if (args === false) {
             return false;
@@ -52,7 +27,7 @@ export default class Restrict extends Command {
         let allChannels = false;
 
         for (let i = 0; i < args.length; i++) {
-            if (args[i].match(/^<#\d+>$/)) channels.push(args[i].slice(2, -1));
+            if (args[i].match(/\<\#\d+\>/g)) channels.push(args[i].slice(2, -1));
             else if (args[i].toLowerCase() == "#{all}") {
                 allChannels = true;
                 channels.push(...msg.guild.channels.filter(channel => channel.type == "text").map(channel => channel.id));
@@ -61,18 +36,22 @@ export default class Restrict extends Command {
                 allCommands = true;
                 commands.push(...client.commands.filter(command => !["restrict", "unrestrict"].includes(command.name)).map(command => command.name));
             }
-            else commands.push(args[i]);
+            else {
+                if (!client.commands.map(cmds => cmds.name).includes(args[i])) continue;
+                if (args[i] === "8ball") commands.push(args[i]);
+                commands.push(args[i]); 
+            }
         }
-    
+
         if (channels.length == 0) {
-            channels = [ msg.channel.id ];
+            channels = [msg.channel.id];
         }
-    
-        if (commands.includes("restrict") || commands.includes("unrestrict")) return await this.fail(msg,"Restricting the restriction commands from being used is probably not what you want to do...");
-    
+
+        if (commands.includes("restrict") || commands.includes("unrestrict")) return await this.fail(msg, "Restricting the restriction commands from being used is probably not what you want to do...");
+
         const serverManager = client.serverManager;
         const server = await serverManager.ensureServer(msg.guild.id);
-    
+
         server.addCommandRestrictions(
             channels.map(channelID => {
                 return {
@@ -81,10 +60,10 @@ export default class Restrict extends Command {
                 };
             })
         );
-    
+
         await server.save();
-    
-        await this.respond(msg, `Restricted in ${allChannels ? "ALL CHANNELS" : channels.map(channelID => `<#${channelID}>`).join(", ")}: \`\`\`${allCommands ? "ALL COMMANDS" : commands.join("\n")}\`\`\``);
+
+        await this.respond(msg, `Restricted in ${allChannels ? "ALL CHANNELS" : channels.map(channelID => `<#${channelID}>`).join(", ")}: \`\`\`${(commands.length > 0) ? (allCommands ? "ALL COMMANDS" : commands.join("\n")) : "None"}\`\`\``);
 
         return true;
     }