+++ src/preload/commands/unrestrict/index.ts
@@ -7,32 +7,7 @@ export default class Unrestrict extends Command {
         super(
             "unrestrict",
             "Unrestricts a command from being used in a channel. Using `{all}` as a command argument unrestrics all commands and using `#{all}` as a channel argument unrestricts the commands in all text channels.",
-            [
-                {
-                    name: "command",
-                    description: "A command to unrestrict. Can be \"{all}\" to unrestrict all commands.",
-                    required: true,
-                    type: "string"
-                },
-                {
-                    name: "command 2",
-                    description: "More commands to unrestrict.",
-                    required: false,
-                    type: "string"
-                },
-                {
-                    name: "channel mention",
-                    description: "The text channel to unrestrict commands in. Can be \"#{all}\" to unrestrict commands in all text channels.",
-                    required: false,
-                    type: "string"
-                },
-                {
-                    name: "channel mention 2",
-                    description: "More text channels to unrestrict commands in.",
-                    required: false,
-                    type: "string"
-                }
-            ],
+            undefined,
             undefined,
             ["SEND_MESSAGES", "MANAGE_CHANNELS"],
             ["text"]
@@ -52,7 +27,7 @@ export default class Unrestrict extends Command {
         let allChannels = false;
 
         for (let i = 0; i < args.length; i++) {
-            if (args[i].match(/^<#\d+>$/)) channels.push(args[i].slice(2, -1));
+            if (args[i].match(/\<\#\d+\>/g)) channels.push(args[i].slice(2, -1));
             else if (args[i].toLowerCase() == "#{all}") {
                 allChannels = true;
                 channels.push(...msg.guild.channels.filter(channel => channel.type == "text").map(channel => channel.id));
@@ -61,7 +36,10 @@ export default class Unrestrict extends Command {
                 allCommands = true;
                 commands.push(...client.commands.filter(command => !["restrict", "unrestrict"].includes(command.name)).map(command => command.name));
             }
-            else commands.push(args[i]);
+            else {
+                if (!client.commands.map(cmds => cmds.name).includes(args[i])) continue;
+                commands.push(args[i]);
+            }
         }
     
         if (channels.length == 0) {
@@ -82,7 +60,7 @@ export default class Unrestrict extends Command {
     
         await server.save();
     
-        await this.respond(msg, `Unrestricted in ${allChannels ? "ALL CHANNELS" : channels.map(channelID => `<#${channelID}>`).join(", ")}: \`\`\`${allCommands ? "ALL COMMANDS" : commands.join("\n")}\`\`\``);
+        await this.respond(msg, `Unrestricted in ${allChannels ? "ALL CHANNELS" : channels.map(channelID => `<#${channelID}>`).join(", ")}: \`\`\`${(commands.length > 0) ? (allCommands ? "ALL COMMANDS" : commands.join("\n")) : "None"}\`\`\``);
 
         return true;
     }